+++ Lab2/Lab2/triangle.h
@@ -5,7 +5,7 @@ using namespace std;
 class Triangle {
 private:
 	int x1, y1, x2, y2, x3, y3;
-	int sideLong_a, sideLong_b, sideLong_c;
+	double sideLong_a, sideLong_b, sideLong_c;
 	int perimeter;
 public: Triangle();
 		Triangle(int x1, int y1, int x2, int y2, int x3, int y3);
@@ -15,7 +15,7 @@ public: Triangle();
 
 		void changeCoordinates();
 		void print(int currentTriangle);
-		int sideLong(int X1, int Y1, int X2, int Y2);
+		double sideLong(int X1, int Y1, int X2, int Y2);
 		void calculations();
 		void checkTriangleType();
 		char* printTriangleType();
